fix sqa sampler
