+++ openjij/sampler/__init__.py
@@ -1,5 +1,6 @@
 from .sampler import *
-from .sa_sampler import *
+from .sa_sampler import SASampler
+from .sqa_sampler import SQASampler
 from .gpu_sampler import *
 from .cmos_annealer import *
 from .response import *
